requirements.txt
absl-py==2.1.0
asttokens==3.0.0
astunparse==1.6.3
attrs==24.3.0
biopython==1.84
biotite==0.41.2
Brotli==1.1.0
captum==0.7.0
certifi==2024.12.14
charset-normalizer==3.4.0
cloudpathlib==0.20.0
cloudpickle==3.1.0
contourpy==1.3.1
cycler==0.12.1
decorator==5.1.1
einops==0.8.0
esm==3.0.0
executing==2.1.0
fair-esm==2.0.0
filelock==3.13.1
flatbuffers==24.12.23
fonttools==4.55.3
fsspec==2024.2.0
gast==0.6.0
google-pasta==0.2.0
grpcio==1.68.1
h5py==3.12.1
huggingface-hub==0.27.0
idna==3.10
ipython==8.30.0
jedi==0.19.2
Jinja2==3.1.3
joblib==1.4.2
keras==3.7.0
kiwisolver==1.4.7
libclang==18.1.1
llvmlite==0.43.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
ml-dtypes==0.4.1
mpmath==1.3.0
msgpack==1.1.0
msgpack-numpy==0.4.8
namex==0.0.8
networkx==3.2.1
numba==0.60.0
numpy==1.26.3
nvidia-cublas-cu11==11.11.3.6
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu11==9.1.0.70
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu11==10.3.0.86
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu11==11.7.5.86
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu11==2.21.5
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu11==11.8.86
nvidia-nvtx-cu12==12.4.127
opt_einsum==3.4.0
optree==0.13.1
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==10.2.0
prompt_toolkit==3.0.48
protobuf==5.29.2
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.18.0
pyparsing==3.2.0
pyteomics==4.7.5
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
safetensors==0.4.5
scikit-learn==1.6.0
scipy==1.14.1
seaborn==0.13.2
shap==0.46.0
six==1.17.0
slicer==0.0.8
stack-data==0.6.3
sympy==1.13.1
tenacity==9.0.0
termcolor==2.5.0
threadpoolctl==3.5.0
tokenizers==0.20.3
torchtext==0.18.0
tqdm==4.67.1
traitlets==5.14.3
transformers==4.46.3
triton==3.1.0
tslearn==0.6.3
typing_extensions==4.9.0
tzdata==2024.2
urllib3==2.2.3
wcwidth==0.2.13
Werkzeug==3.1.3
wrapt==1.17.0
zstd==1.5.6.1
